add test for generatingrandm bean with fluent setter
