+++ easy-random-core/src/test/java/org/jeasy/random/beans/FluentSetterBean.java
@@ -1,0 +1,15 @@
+package org.jeasy.random.beans;
+
+public class FluentSetterBean {
+
+    public String name;
+
+    public FluentSetterBean setName(String name) {
+        this.name = name;
+        return this;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
