+++ easy-random-core/src/test/java/org/jeasy/random/util/ReflectionUtilsTest.java
@@ -298,17 +298,4 @@ class ReflectionUtilsTest {
 
     }
 
-    private class FluentSetterBean {
-
-        public String name;
-
-        public FluentSetterBean setName(String name) {
-            this.name = name;
-            return this;
-        }
-
-        public String getName() {
-            return name;
-        }
-    }
 }
